generalize the overrides types as well
